same for material/materialstack
